Create a template to support for Horizontal Pod Autoscaling (HPA) (#695)
* docs(hpa): hpa-values.yaml is an example on how to activate HPA

The hpa-values.yaml file includes an example on how to activate the HPA configuration for the runtime-bundle microservice.
In production the values must be changed according to the actual CPU and Memory values.

Refs: Activiti/Activiti#3791

* docs(ingress): deleted the nginx-ingress version not supported anymore

Ref: Activiti/Activiti#3791

* docs(HPA): added documentation on the new HPA feature

Adding documentation on how to enable and use the HorizontalPodAutoscaler
Explaining the corner cases for the query consumer where the HPA can't be used

Refs: Activiti/Activiti#3791

* fix(hpa): increasing the thresholds to trigger HPA

After some more tests, I increased the thresholds to trigger HPA since the runtime bundle is quite demanding in terms of resources but still able to withstand a fair amount of requets

Ref: Activiti/Activiti#3791

* chore(hpa): adding an example for query in hpa-values.yaml file

Ref: Activiti/Activiti#3791

* chore(hpa): setting partition number to 4 in the example file

Ref: Activiti/Activiti#3791

* docs(hpa): aligning the README.md with provided values.yaml files

Ref: Activiti/Activiti#3791

* docs(hpa): documented scaling policies

Ref: Activiti/Activiti#3791

* docs(hpa): rephrasing and refining the query consumers docs

Refs: Activiti/Activiti#3791

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>